[KA-123] feat(rag): Populating insights.

+++ src/Kentico.Xperience.AiraUnified/Insights/ContactGroupModel.cs
@@ -4,6 +4,6 @@
     {
         public string Name { get; set; } = string.Empty;
         public int ContactCount { get; set; }
-        public double RatioPercentage { get; set; }
+        public decimal RatioPercentage { get; set; }
     }
 }
